requirements.txt
alembic==1.13.2
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
httptools==0.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
watchfiles==0.23.0
websockets==12.0
SQLAlchemy==2.0.32
